WIP 0.1.0
- Проверка на наличие папки

+++ ACP/backup.cpp
@@ -36,28 +36,45 @@ namespace backup
 
     void Back::patchAll(const fs::path path)
     {
-        for (const auto& entry : fs::directory_iterator(path))
+        if (fs::exists(path))
         {
 
-            std::cout << "\t" << entry.path().stem() << std::endl;
+            for (const auto& entry : fs::directory_iterator(path))
+            {
 
+                std::cout << "\t" << entry.path().stem() << std::endl;
+
+            }
+            std::cout << "____________________________\n\n";
         }
-        std::cout << "____________________________\n\n";
+        else 
+        {
+
+            std::cerr << "\t\tОшибка: " << path << " не существует\n";
+        }
+        
     }
 
     void Back::patchConcrect(const fs::path path, const std::unordered_set<std::string> allowedNames)
     {
-        for (const auto& entry : fs::directory_iterator(path))
-        {
-            std::string filename = entry.path().filename().string();
-
-            if (allowedNames.count(filename) > 0)
+        if (fs::exists(path))
+        { 
+            for (const auto& entry : fs::directory_iterator(path))
             {
-                std::cout << "\t" << entry.path().stem() << std::endl;
-            }
+                std::string filename = entry.path().filename().string();
 
+                if (allowedNames.count(filename) > 0)
+                {
+                    std::cout << "\t" << entry.path().stem() << std::endl;
+                }
+
+            }
+            std::cout << "____________________________\n\n";
         }
-        std::cout << "____________________________\n\n";
+        else
+        {
+            std::cerr << "\t\tОшибка: " << path << " не существует\n";
+        }
     }
 
     void Back::input(const fs::path path, const std::string folder_name)
@@ -69,7 +86,7 @@ namespace backup
         std::cout << "2 - Просмотреть все папки, которые соответсвуют фильтру\n";
         std::cout << "3 - Сделать бекап\n";
         std::cout << "4 - Изменить основную папку\n";
-        std::cout << "5 - Изменить имя создаваеой папки\n";
+        std::cout << "5 - Изменить имя создаваемой папки\n";
         std::cout << "\n>>";
 
     }
@@ -91,22 +108,29 @@ namespace backup
         // Копируем рекурсивно все содержимое
 
         // Папка
-        try {
-            std::string folder_name1 = origname;
-            fs::path sourcefolder = destination.string() + "/" + folder_name1;
-            // Путь
-            fs::path dir_path = destination / folder_name1;
-            fs::create_directory(dir_path);
+        if (fs::exists(source))
+        { 
+            try {
+                std::string folder_name1 = origname;
+                fs::path sourcefolder = destination.string() + "/" + folder_name1;
+                // Путь
+                fs::path dir_path = destination / folder_name1;
+                fs::create_directory(dir_path);
 
-            //std::cout << "\tПапка в: " << sourcefolder.string() << " С именем " << folder_name ;
-            //std::cout << "\n\tСоздаю " << source << " --- " << destination << " \n ";
+                //std::cout << "\tПапка в: " << sourcefolder.string() << " С именем " << folder_name ;
+                //std::cout << "\n\tСоздаю " << source << " --- " << destination << " \n ";
 
-            fs::copy(source, sourcefolder, fs::copy_options::recursive | fs::copy_options::overwrite_existing);
-            std::cout << "\t\tПапка скопирована: " << source << " -> " << sourcefolder << "\n\n";
+                fs::copy(source, sourcefolder, fs::copy_options::recursive | fs::copy_options::overwrite_existing);
+                std::cout << "\t\tПапка скопирована: " << source << " -> " << sourcefolder << "\n\n";
+            }
+            catch (const fs::filesystem_error& e) {
+               // std::cerr << "Ошибка копирования: " << e.what() << std::endl;
+
+            }
         }
-        catch (const fs::filesystem_error& e) {
-           // std::cerr << "Ошибка копирования: " << e.what() << std::endl;
-
+        else
+        {
+            std::cerr << "\t\tОшибка: " << source << " не существует\n";
         }
 
     }
@@ -114,24 +138,31 @@ namespace backup
     void Back::createFolder(const fs::path& path, std::string folder_name, std::unordered_set<std::string> allowedNames)
     {
         // Папка
-        std::string folder_name1 = folder_name + time();
-        // Путь
-        fs::path dir_path = path / folder_name1;
-        fs::create_directory(dir_path);
-        std::cout << "\n>> Папка создана! " << dir_path;
-        std::cout << "\nКопируем... \n";
-        std::string filepathEnd = path.string() + "/" + folder_name1;
-        for (const auto& entry : fs::recursive_directory_iterator(path))
+        if (fs::exists(path))
         {
-            std::string filename = entry.path().filename().string();
-            std::string filepath = entry.path().string();
-
-            if (allowedNames.count(filename) > 0)
+            std::string folder_name1 = folder_name + time();
+            // Путь
+            fs::path dir_path = path / folder_name1;
+            fs::create_directory(dir_path);
+            std::cout << "\n>> Папка создана! " << dir_path;
+            std::cout << "\nКопируем... \n";
+            std::string filepathEnd = path.string() + "/" + folder_name1;
+            for (const auto& entry : fs::recursive_directory_iterator(path))
             {
-                //std::cout << "\n\tСоздаю " << filepath << " === " << filepathEnd << " \n ";
-                copyDirectory(filepath, filepathEnd, filename);
-                //std::cout << "\tПапка с путем " << filepath << " Скопирована в " << filepathEnd << " C именем " << filename << "\n";
+                std::string filename = entry.path().filename().string();
+                std::string filepath = entry.path().string();
+
+                if (allowedNames.count(filename) > 0)
+                {
+                    //std::cout << "\n\tСоздаю " << filepath << " === " << filepathEnd << " \n ";
+                    copyDirectory(filepath, filepathEnd, filename);
+                    //std::cout << "\tПапка с путем " << filepath << " Скопирована в " << filepathEnd << " C именем " << filename << "\n";
+                }
             }
+        }
+        else
+        {
+            std::cerr << "\t\tОшибка: " << path << " не существует\n";
         }
     }
    
@@ -141,7 +172,9 @@ namespace backup
         std::string fullpath = "explorer " + path;
         std::replace(fullpath.begin(), fullpath.end(), '/', '\\');
         system(fullpath.c_str());
-        std::cout << fullpath;
+
 
     }
+
+
 }
